requirements.txt
Django==1.6.5
South==0.8.4
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.18
celery[redis]==3.1.12
celerymon==1.0.3
django-celery==3.1.10
django-extensions==1.3.8
gnureadline==6.3.3
ipython==2.1.0
kombu==3.0.20
pytz==2014.4
redis==2.10.1
six==1.7.3
tornado==3.2.2
raven==5.0.0
djangorestframework==2.3.14
Markdown==2.4.1
-e git+https://github.com/jdunck/python-unicodecsv.git@13a300d82d8002e1db141540830e10f3c8dd7606#egg=unicodecsv-master
